// src/interfaces/Imovel/comissao.tsx
type comissaoEnvolvidos = {
  cpf_cnpj?: string
  empresa_cnpj?: string
  agencia?: string
  banco_id?: string
  nome_banco?: string
  cnpj?: string
  cpf?: string
  creci?: string
  desconto?: string
  id: string | number
  name?: string
  nome_empresarial?: string
  numero_conta?: string
  pix?: string
  chave_pix?: "CPF ou CNPJ" | "Celular" | "E-mail" | "Chave aleatória"
  porcentagem_comissao?: string
  porcentagem_real?: string
  tipo_pagamento?: "pix" | "banco"
  tipo_pessoa?: "PF" | "PJ"
  usuario_id?: string | number
  valor_real?: string
  tipo_laudo_opcionista?: "simples" | "com_chave" | "exclusividade" | "lançamento"
  tipo_laudo_opcionista_label?: "Simples" | "Com chave" | "Exclusividade" | "Lançamentos" | "Tipo não cadastrado"
  total_comissao_corretor?: string
  tipo_chave_pix_id?: string
  value_autocomplete?: { usuario_id: number, label: string }
}
// ...
type dadosRepasseType = {
  valor_benesh_comunicacao: string | number
  agencia: string | null
  banco_id: number | null
  cnpj?: string
  comissao_id: number
  company_bank_details_id: number
  created_at: string
  creci_empresa: string
  desconto: string | ""
  empresa_cnpj: string
  empresa_id: null
  empresa_resse_id: number
  id: number
  nome_banco: string | null
  nome_empresarial: string
  numero_conta: string | null
  pix: string
  porcentagem: string
  processo_id: number
  repasse_franquia: null
  tipo_pagamento: string
  updated_at: string
  valor_real: string
}